pyproject.toml
[tool.poetry]
name = "filter-spotify-playlists"
version = "0.1.0"
description = ""
authors = ["Alejandro Cuadrado Torre <alejandro.cuadrado.torre@gmail.com>"]
readme = "README.md"
packages = [{ include = "filter_spotify_playlists" }]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
uvicorn = { extras = ["standard"], version = "^0.23.2" }
pydantic-settings = "^2.0.3"
spotipy = "^2.23.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
